chore: removed unused files
