+++ web/components/chat/ChatInputSetting.tsx
@@ -17,7 +17,7 @@ import {
 	PopoverContent,
 	PopoverTrigger,
 } from '@components/ui/popover';
-import useChatStore from '@context/chat.store';
+import useChatSessions from '@context/chatSessions.store';
 import { useUIStore } from '@context/ui.store';
 import { useUser } from '@context/user';
 import useDocumentTabs, { TabType } from '@hooks/useDocumentTabs';
@@ -25,70 +25,52 @@ import useGetEditorRouter from '@hooks/useGetEditorRouter';
 import { useGetUserUploads } from '@resources/editor-page';
 import clsx from 'clsx';
 import { BookUp, Globe2, Library, RefreshCw, Search } from 'lucide-react';
-import React, { ChangeEvent, memo, useCallback } from 'react';
+import React, { memo } from 'react';
 import { FaStopCircle } from 'react-icons/fa';
 
 type Props = {
-	cancelHandler: () => void;
-	regenerateHandler: () => void;
-	isRegenerateSeen?: boolean;
-	searchInputValue: string;
-	setSearchInputValue: React.Dispatch<React.SetStateAction<string>>;
-	isLoading: boolean;
-	isSearching: boolean;
-	isSearchOpen: boolean;
-	setIsSearchOpen: React.Dispatch<React.SetStateAction<boolean>>;
-	isSettingsOpen: boolean;
-	setIsSettingsOpen: React.Dispatch<React.SetStateAction<boolean>>;
+	sessionId: string;
+	minimized?: boolean;
 };
 
-const ChatInputSetting = ({
-	regenerateHandler,
-	isRegenerateSeen,
-	cancelHandler,
-	searchInputValue,
-	setSearchInputValue,
-}: Props) => {
+const ChatInputSetting = ({ sessionId, minimized }: Props) => {
 	const { user } = useUser();
+	const chatContext = useChatSessions(
+		s => s.chatSessions[sessionId]?.chatContext,
+	);
+	const searchInput = useChatSessions(
+		s => s.chatSessions[sessionId]?.chatSearchInput,
+	);
+	const isHandling = useChatSessions(
+		s => s.chatSessions[sessionId]?.isHandling,
+	);
+	const activeFileReference = useChatSessions(
+		s => s.chatSessions[sessionId]?.activeFileReference,
+	);
+	const { setChatContext, setChatSearchInput, setActiveFileReference } =
+		useChatSessions(s => s.actions);
+
 	const { projectId } = useGetEditorRouter();
-	const chatContext = useChatStore(s => s.chatContext);
-	const setChatContext = useChatStore(s => s.setChatContext);
-
-	const isHandling = useChatStore(s => s.isHandling);
-
-	const setActiveFileReference = useChatStore(s => s.setActiveFileReference);
 	const { data: uploadFiles } = useGetUserUploads(user?.id, projectId);
-	const activeFileReference = useChatStore(s => s.activeFileReference);
 	const { openDocument } = useDocumentTabs();
 
 	const rightPanelWidth = useUIStore(s => s.rightPanelWidth);
-
-	const onChange = useCallback((e: ChangeEvent) => {
-		const value = (e.target as HTMLInputElement).value;
-		if (value) {
-			setSearchInputValue(value);
-		}
-	}, []);
+	const isRegenerateSeen = false; // TODO
 
 	return (
 		<div className="flex justify-between pt-2 pb-2 text-xs">
 			<div className="flex  max-w-full gap-2  ">
 				<DropdownMenu>
 					<DropdownMenuTrigger>
-						<div
-							onClick={() => {
-								// setIsSettingsOpen(true);
-							}}
-							className="px-2 py-1 transition-all duration-100 ease-in-out bg-white rounded-md border hover:bg-gray-100 dark:hover:bg-neutral-800 cursor-pointer dark:bg-neutral-950 "
-						>
+						<div className="px-2 py-1 transition-all duration-100 ease-in-out bg-white rounded-md border hover:bg-gray-100 dark:hover:bg-neutral-800 cursor-pointer dark:bg-neutral-950 ">
 							Context:{' '}
-							{chatContext === 'project'
-								? 'Default'
+							{chatContext === 'file'
+								? 'PDF'
 								: chatContext === 'realtime'
 								? 'Realtime'
 								: chatContext === 'references'
 								? activeFileReference?.name || 'All Documents'
-								: 'PDF'}
+								: 'Default'}
 						</div>
 					</DropdownMenuTrigger>
 					<DropdownMenuContent
@@ -101,7 +83,7 @@ const ChatInputSetting = ({
 							Chat Context
 						</DropdownMenuLabel>
 						<DropdownMenuItem
-							onClick={() => setChatContext('project')}
+							onClick={() => setChatContext(sessionId, 'project')}
 							className="py-2"
 						>
 							<Logomark className="h-8 w-8 mr-2" />
@@ -134,8 +116,8 @@ const ChatInputSetting = ({
 														'bg-neutral-100 font-semibold dark:bg-neutral-800',
 												)}
 												onClick={() => {
-													setChatContext('references');
-													setActiveFileReference({
+													setChatContext(sessionId, 'references');
+													setActiveFileReference(sessionId, {
 														name: null,
 														fileId: null,
 													});
@@ -161,17 +143,18 @@ const ChatInputSetting = ({
 													onClick={() => {
 														const name =
 															file.custom_citation?.title || file.file_name;
-														setChatContext('references');
-														setActiveFileReference({
+														setChatContext(sessionId, 'references');
+														setActiveFileReference(sessionId, {
 															name:
 																file.custom_citation?.title || file.file_name,
 															fileId: file.id,
 														});
-														openDocument({
-															source: file.id,
-															label: name,
-															type: TabType.UserUpload,
-														});
+														minimized &&
+															openDocument({
+																source: file.id,
+																label: name,
+																type: TabType.UserUpload,
+															});
 													}}
 												>
 													<div>
@@ -191,8 +174,8 @@ const ChatInputSetting = ({
 											key="all-documents"
 											disabled
 											onClick={() => {
-												setChatContext('references');
-												setActiveFileReference({
+												setChatContext(sessionId, 'references');
+												setActiveFileReference(sessionId, {
 													name: null,
 													fileId: null,
 												});
@@ -206,7 +189,7 @@ const ChatInputSetting = ({
 							</DropdownMenuPortal>
 						</DropdownMenuSub>
 						<DropdownMenuItem
-							onClick={() => setChatContext('realtime')}
+							onClick={() => setChatContext(sessionId, 'realtime')}
 							className="py-2"
 						>
 							<Globe2 className="h-5 w-5 mr-3.5 ml-1.5" />
@@ -219,91 +202,58 @@ const ChatInputSetting = ({
 						</DropdownMenuItem>
 					</DropdownMenuContent>
 				</DropdownMenu>
-				{/* //TODO: Creativity doesnt do anything right now */}
-				{/* <Popover>
-					<PopoverTrigger>
-						<div
-							onClick={() => {
-								// setIsSettingsOpen(true);
-							}}
-							className="px-2 py-1 transition-all duration-100 ease-in-out bg-white rounded-md border hover:bg-gray-100 dark:hover:bg-neutral-800 cursor-pointer dark:bg-neutral-950 "
-						>
-							Creativity: {(temperature * 100).toFixed(0)} %
-						</div>
-					</PopoverTrigger>
-					<PopoverContent side="top" sideOffset={10}>
-						<div className="mb-4">
-							<div className="font-medium">Creativity</div>
-							<div className="text-xs text-neutral-500">
-								Lower creativity means generating answers based on the context
-								strictly.
-							</div>
-						</div>
-						<Slider
-							className="bg-transparent"
-							min={0}
-							max={1}
-							step={0.01}
-							defaultValue={[temperature]}
-							onValueChange={value => {
-								setTemperature(value[0]);
-							}}
-							id="creativity"
-						/>
-						<div className="flex items-center justify-between text-sm text-neutral-500">
-							<span className="text-xs">0</span>
-							<span className="text-xs">100</span>
-						</div>
-					</PopoverContent>
-				</Popover> */}
+
+				{/* // TODO: Figure out how to stop generation */}
 				{isHandling && (
-					<InputSettingButton onClick={cancelHandler}>
+					<InputSettingButton onClick={() => alert('cancel')}>
 						Stop Generation{' '}
 						<FaStopCircle size="10" className="inline-block animate-pulse" />
 					</InputSettingButton>
 				)}
 				{isRegenerateSeen && (
-					<InputSettingButton onClick={regenerateHandler}>
+					<InputSettingButton onClick={() => alert('regemerate')}>
 						Regenerate <RefreshCw size="10" className="inline-block" />
 					</InputSettingButton>
 				)}
 			</div>
 			{!isHandling && (
-				<Popover>
-					<PopoverTrigger asChild>
-						<div className="flex text-xs items-center gap-1 px-2 py-1 transition-all duration-100 ease-in-out bg-white rounded-md border hover:bg-gray-100 dark:hover:bg-neutral-800 cursor-pointer dark:bg-neutral-950">
-							<Search size="10" className="inline-block" /> Search
-						</div>
-					</PopoverTrigger>
-					<PopoverContent
-						side="top"
-						align="end"
-						className="p-1"
-						style={{ width: rightPanelWidth - 46 }}
-					>
-						<div
-							onSubmit={e => {
-								e.preventDefault();
-								alert('search');
-							}}
-							className="text-xs text-neutral-500 relative flex max-h-max"
+				<>
+					<Popover>
+						<PopoverTrigger asChild>
+							<div className="flex text-xs items-center gap-1 px-2 py-1 transition-all duration-100 ease-in-out bg-white rounded-md border hover:bg-gray-100 dark:hover:bg-neutral-800 cursor-pointer dark:bg-neutral-950">
+								<Search size="10" className="inline-block" /> Search
+							</div>
+						</PopoverTrigger>
+						<PopoverContent
+							side="top"
+							align="end"
+							className="p-1"
+							style={{ width: rightPanelWidth - 46 }}
 						>
-							<Search
-								strokeWidth={1.4}
-								className="absolute my-3 ml-2 h-4 w-4"
-							/>
-							<Input
-								className="w-full focus-visible:ring-0 focus-visible:ring-ring focus-visible:ring-offset-0 pl-10"
-								value={searchInputValue}
-								onChange={onChange}
-								onKeyDown={e => {
-									// Prevent Enter from triggering form submit
-									if (e.key === 'Enter') e.preventDefault();
+							<div
+								onSubmit={e => {
+									e.preventDefault();
+									alert('search');
 								}}
-							/>
-						</div>
-					</PopoverContent>
-				</Popover>
+								className="text-xs text-neutral-500 relative flex max-h-max"
+							>
+								<Search
+									strokeWidth={1.4}
+									className="absolute my-3 ml-2 h-4 w-4"
+								/>
+								<Input
+									className="w-full focus-visible:ring-0 focus-visible:ring-ring focus-visible:ring-offset-0 pl-10"
+									value={searchInput}
+									onChange={e => setChatSearchInput(sessionId, e.target.value)}
+									onKeyDown={e => {
+										// Prevent Enter from triggering form submit
+										if (e.key === 'Enter') e.preventDefault();
+									}}
+								/>
+							</div>
+						</PopoverContent>
+					</Popover>
+				</>
 			)}
 		</div>
 	);
