+++ web/components/chat/ChatPanel.tsx
@@ -1,6 +1,6 @@
-import ChatboxV2 from '@components/chatV2/ChatboxV2';
-import ChatInputSetting from '@components/chatV2/ChatInputSettingV2';
-import ChatInput from '@components/chatV2/ChatInputV2';
+import Chatbox from '@components/chat/Chatbox';
+import ChatInputSetting from '@components/chat/ChatInputSetting';
+import ChatInput from '@components/chat/ChatInput';
 import useChatSessions from '@context/chatSessions.store';
 import { retrieveChatMessages } from '@hooks/api/isaac/useRetrieveChatMessages';
 import { UniqueTabSources } from '@hooks/useDocumentTabs';
@@ -29,7 +29,7 @@ const ChatPanel = () => {
 
 	return (
 		<>
-			<ChatboxV2 sessionId={sessionId} minimized />
+			<Chatbox sessionId={sessionId} minimized />
 			<div className="absolute w-full bottom-0 px-3">
 				<ChatInputSetting minimized sessionId={sessionId} />
 				<ChatInput minimized sessionId={sessionId} />
