+++ web/components/chat/ChatSessionCard.tsx
@@ -3,10 +3,9 @@ import {
 	ArrowUpRight,
 	MessageSquareIcon,
 	MessageSquarePlusIcon,
-	ScrollTextIcon,
 	X,
 } from 'lucide-react';
-import React, { useState } from 'react';
+import React from 'react';
 import { ChatSessionType } from 'types/chat';
 
 export default function ChatSessionCard({
