+++ web/pages/api/chat.ts
@@ -55,6 +55,8 @@ export default async function (req: NextApiRequest, res) {
 		if (context === 'references') {
 			const body = uploadId ? { prompt, uploadId } : { prompt, projectId };
 
+			console.log(body)
+
 			const res = await fetch(
 				uploadId ? singleReferenceEndpoint : projectReferenceEndpoint,
 				{
