+++ web/pages/editor/[id].tsx
@@ -1,6 +1,6 @@
 // import ReferenceViewer from '@components/core/ReferenceViewer';
 import Chat from '@components/chat/Chat';
-import ChatTab from '@components/chatV2/ChatTab';
+import ChatTab from '@components/chat/ChatTab';
 import EditorEmptyState from '@components/editor/EditorEmptyState';
 import AppLayout from '@components/editor/EditorLayout';
 import LiteratureSearchTab from '@components/literature/LiteratureSearchTab';
